Fixed some bugs related to post deletion Fixed a bug with the popup templates not displaying the title Added a check to make sure $badimgs in common.php's delimgs is non-null before attempting to iterate through it

+++ branches/drydock-june11-SQLITE-TEST-BUILD/common.php
@@ -347,22 +347,25 @@
 	 */
 	function delimgs($badimgs)
 	{
-		//Delete these images
-		foreach ($badimgs as $bad)
+		if($badimgs != null)
 		{
-			$bad=(int)$bad;
-			if($bad!=0)
+			//Delete these images
+			foreach ($badimgs as $bad)
 			{
-				$pyath=THpath."images/".$bad."/";
-				$it=opendir($pyath);
-				while (($img=readdir($it))!==false)
+				$bad=(int)$bad;
+				if($bad!=0)
 				{
-					if ($img{0}!=".")
+					$pyath=THpath."images/".$bad."/";
+					$it=opendir($pyath);
+					while (($img=readdir($it))!==false)
 					{
-						unlink($pyath.$img);
+						if ($img{0}!=".")
+						{
+							unlink($pyath.$img);
+						}
 					}
+					rmdir($pyath);
 				}
-				rmdir($pyath);
 			}
 		}
 	}
